refactor: LadderResult의 패키지를 변경한다
- ladder_result -> ladder.result로 수정하였습니다.
